Separate some server side ComponentStore
