+++ src/mrtjp/projectred/integration/ComponentStore.java
@@ -24,7 +24,11 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-public class ComponentStore
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+
+@SideOnly(Side.CLIENT)
+public class ComponentStore extends ComponentStoreServer
 {
     public static CCModel base = loadBase("base");
     public static CCModel lightChip = loadModel("chip");
@@ -164,20 +168,6 @@ public class ComponentStore
         return m;
     }
 
-    public static Transformation orientT(int orient)
-    {
-        Transformation t = Rotation.sideOrientation(orient%24>>2, orient&3);
-        if (orient >= 24)
-            t = new Scale(-1, 1, 1).with(t);
-
-        return t.at(Vector3.center);
-    }
-
-    public static Transformation dynamicT(int orient)
-    {
-        return orient == 0 ? new RedundantTransformation() : new Scale(-1, 1, 1).at(Vector3.center);
-    }
-
     public static CCModel bakeCopy(CCModel base, int orient)
     {
         CCModel m = base.copy();
@@ -206,28 +196,6 @@ public class ComponentStore
         }
         return m;
     }
-
-    public static IndexedCuboid6[] buildCubeArray(int xSize, int zSize, Cuboid6 box, Vector3 expand)
-    {
-        Vector3 min = box.min;
-        Vector3 max = box.max;
-        min.multiply(1/16D);
-        max.multiply(1/16D);
-        expand.multiply(1/16D);
-        IndexedCuboid6[] data = new IndexedCuboid6[xSize*zSize];
-        for (int i = 0; i < data.length; i++)
-        {
-            int x = i%xSize;
-            int z = i/zSize;
-            double dx = (max.x-min.x)/xSize;
-            double dz = (max.z-min.z)/zSize;
-            Vector3 min1 = new Vector3(min.x+dx*x, min.y, min.z+dz*z);
-            Vector3 max1 = new Vector3(min1.x+dx, max.y, min1.z+dz);
-            data[i] = new IndexedCuboid6(i, new Cuboid6(min1, max1).expand(expand));
-        }
-        return data;
-    }
-
 
     public static abstract class ComponentModel
     {
